Initial build scripts
You should build

    gcc -o run run.c

for a workaround on a bug.

Also, these scripts don't attempt to clone anything.  However, they do
build everything under '_build', and put the generated image files in
'out'.

In addition, 'term.rkt' will monitor and re-open picocom to handle cases
where the uart device goes away (say when hard resetting the board).

'flash.rkt' manually mounts the device, copies the build file over, and
unmounts it.  It should be able to find the device, as long as systemd
or udev is working, and as long as there is only a single daplink device
present.